requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
python-binance
torch
pytest

